Classifier results for different genes (#52)
* Classifier results for different genes

Hi all,
Since MVP is going to be released soon and open to the public, maybe it would be useful to provide some kind of demo to explain how the package works. It would be like providing tutorial/examples for the first-time users. 
Also, since most likely sklearn.linear_model.SGDClassifier is going to be used as a classifier, it would be nice to see the results for different genes, not only TP53. During the last meeting there was a discussion regarding this. Since it would take really long time to run the code for all genes, maybe it's better to focus on the group of genes that are considered dominant by a certain criteria. 
I have submitted pull requests for two files. Both files output AUROCs for a list of genes. First file is a code that determines AUROCs for a list of tumor suppressor genes. The list is not exhaustive and it is copied from https://www.cancerquest.org/cancer-biology/cancer-genes/tumor-suppressors/tumor-suppressor-table. The second code runs the cla

* Delete 2.TCGA-MLexample-LungCancer.ipynb

* Delete 2.TCGA-MLexample-LungCancer.py

* Delete 2.TCGA-MLexample-TumorSuppressors.ipynb

* Delete 2.TCGA-MLexample-TumorSuppressors.py

* Adding one notebook with two plots

Two plots are generated - a lung cancer gene plot and a suppressor gene plot

* Adding .py file

* Delete .py file

* Delete .ipynb file

* Adding one notebook with two plots

* Delete Classifier_results-different_genes.ipynb

* Delete Classifier_results-different_genes.py

* Notebook with implemented two changes

Definition of functions outside the loop
Append command instead of +=

* Delete Classifier_results-different_genes.ipynb

* Delete Classifier_results-different_genes.py

* Results for different set of genes

This file contains results for:
-Lung Cancer genes
-Tumor suppressors
-Oncogenes

Oncogenes are chosen based on the categories listed on https://en.wikipedia.org/wiki/Oncogene#Classification.
The required changes are incorporated as well as two other things:
1. The get_threshold_metrics() function is removed since the code only plots AUROCs. The AUROC value is directly computed using roc_auc_score().
2. The suppression of  joblib warning is added at the end of the for loop code. For some reason, I get a warning regarding the 'the shape of non-C contiguous array'.

* Delete Classifier_results-different_genes.ipynb

* Update Classifier_results-different_genes.py

* Notebook with updated gene symbols